move srcs to src/
