+++ locale.ts
@@ -1,6 +1,6 @@
 import { useRouter } from "next/router";
-import en from "./locales/en";
-import ja from "./locales/ja";
+import en from "./src/locales/en";
+import ja from "./src/locales/ja";
 
 export const useLocale = () => {
   const { locale } = useRouter();
