+++ src/components/molecules/sign_out_button.tsx
@@ -1,9 +1,9 @@
 import { Box, HStack, Text } from "@chakra-ui/react";
 import Link from "next/link";
 import { useState } from "react";
-import { auth } from "../../firebaseConfig";
+import { auth } from "../../../firebaseConfig";
 import { MdLogout } from "react-icons/md";
-import { useLocale } from "../../locale";
+import { useLocale } from "../../../locale";
 
 export default function SignOutButton() {
   const [loading, setLoading] = useState(false);
