+++ src/components/organisms/header.tsx
@@ -16,10 +16,10 @@ import {
   Text,
 } from "@chakra-ui/react";
 import { useAuthState } from "react-firebase-hooks/auth";
-import { auth } from "../../firebaseConfig";
+import { auth } from "../../../firebaseConfig";
 import SignOutButton from "../molecules/sign_out_button";
 import { useRouter } from "next/router";
-import { useLocale } from "../../locale";
+import { useLocale } from "../../../locale";
 import { MdAdd, MdExpandMore } from "react-icons/md";
 
 export default function Header() {
