+++ src/components/organisms/new_event_form.tsx
@@ -19,13 +19,13 @@ import {
 import { User as FirebaseUser } from "firebase/auth";
 import { useRouter } from "next/router";
 import { useState } from "react";
-import { db } from "../../firebaseConfig";
+import { db } from "../../../firebaseConfig";
 import { Event, eventConverter } from "../../types/event";
 import { User, userConverter } from "../../types/user";
 import Loading from "../atoms/loading";
 import UserTag from "../atoms/user_tag";
 import { MdPersonAddAlt1, MdSaveAlt } from "react-icons/md";
-import { useLocale } from "../../locale";
+import { useLocale } from "../../../locale";
 
 type NewEventFormProps = {
   firebaseUser: FirebaseUser;
