+++ src/locales/en.ts
