+++ src/locales/ja.ts
