+++ src/pages/new.tsx
@@ -6,8 +6,8 @@ import router from "next/router";
 import { useAuthState } from "react-firebase-hooks/auth";
 import Loading from "../components/atoms/loading";
 import NewEventForm from "../components/organisms/new_event_form";
-import { auth } from "../firebaseConfig";
-import { useLocale } from "../locale";
+import { auth } from "../../firebaseConfig";
+import { useLocale } from "../../locale";
 
 export default function NewEvent() {
   const [user, loadingUser] = useAuthState(auth);
