+++ src/pages/p/[id].tsx
@@ -27,8 +27,8 @@ import { MdArrowBack } from "react-icons/md";
 import Loading from "../../components/atoms/loading";
 import NoItems from "../../components/atoms/no_items";
 import { SomethingWentWrong } from "../../components/atoms/something_went_wrong";
-import { auth, db } from "../../firebaseConfig";
-import { useLocale } from "../../locale";
+import { auth, db } from "../../../firebaseConfig";
+import { useLocale } from "../../../locale";
 import { paymentMethodConverter } from "../../types/payment_method";
 import { userConverter } from "../../types/user";
